refactor - #47 공통 constants 파일 분리
