+++ src/main/java/sopt/org/HMH/domain/user/domain/UserConstants.java
@@ -1,0 +1,5 @@
+package sopt.org.HMH.domain.user.domain;
+
+public abstract class UserConstants {
+    public static final Integer INITIAL_POINT = 0;
+}
